transition to multi-page dash config
